pyproject.toml
[tool.poetry]
name = "smartfactory-process-graph-optimization"
version = "0.1.0"
description = ""
authors = ["Angel Fernandez Sobrino <fernandezsobrinoangel@gmail.com>"]
readme = "README.md"
packages = [{include = "smartfactory_process_graph_optimization"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
pyyaml = "^6.0.1"
prettytable = "^3.9.0"
pyvis = "^0.3.2"
flask = "^3.0.2"
pyvisgraph = "^0.2.1"
tqdm = "^4.66.2"
cython = "^3.0.9"
svg-py = "^1.4.3"

[tool.black]
line-length = 88

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
types-pyyaml = "^6.0.12.12"
black = "^23.12.1"
pylint = "^3.0.3"
poethepoet = "^0.24.4"
ipykernel = "^6.29.0"
pyinstaller = "^6.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poe.tasks]
graph = "python graph.py"
main = "python src/main.py"
serve = "python src/server.py"
generate_exe = "pyinstaller -w -F --add-data 'src/static:static' -p 'src' src/server.py"
export_requirements = { "shell" = "poetry export --without-hashes --format=requirements.txt > requirements.txt" }
build = ["generate_exe", "export_requirements"]
